// src/app/interfaces/Empresa/IEmpresa.ts
export interface IEmpresa {
    id?: number
    nombre?: string
    direccion?: string
    telefono?: string
    email?: string
    redes_sociales?: string
    logo?: string
    lema?: string
    user_crea?: string
    user_actualiza?: string
    user_elimina?: string
    sistema?: boolean
    estado?: boolean
}

export interface EmpresaResponse {
    result: boolean
    message?: string
    data?: IEmpresa | IEmpresa[]
    error?: string
    status?: number
}

export interface IEmpresaPaginate {
    currentPage: number
    limit: number
    totalPages: number
    totalItems: number
    nextPage: number | null
    previousPage: number | null
}

export interface EmpresaResponsePaginate {
    result: boolean
    message?: string
    data?: IEmpresa[]
    pagination?: IEmpresaPaginate
    error?: string
    status?: number
}